First commit of module_utils shared code
