+++ plugins/modules/mongodb_status.py
@@ -136,51 +136,7 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.six import binary_type, text_type
 from ansible.module_utils.six.moves import configparser
 from ansible.module_utils._text import to_native
-
-
-# =========================================
-# MongoDB module specific support methods.
-#
-
-def check_compatibility(module, srv_version, driver_version):
-    """Check the compatibility between the driver and the database.
-
-    See: https://docs.mongodb.com/ecosystem/drivers/driver-compatibility-reference/#python-driver-compatibility
-
-    Args:
-        module: Ansible module.
-        srv_version (LooseVersion): MongoDB server version.
-        driver_version (LooseVersion): Pymongo version.
-    """
-    msg = 'pymongo driver version and MongoDB version are incompatible: '
-
-    if srv_version >= LooseVersion('4.2') and driver_version < LooseVersion('3.9'):
-        msg += 'you must use pymongo 3.9+ with MongoDB >= 4.2'
-        module.fail_json(msg=msg)
-
-    elif srv_version >= LooseVersion('4.0') and driver_version < LooseVersion('3.7'):
-        msg += 'you must use pymongo 3.7+ with MongoDB >= 4.0'
-        module.fail_json(msg=msg)
-
-    elif srv_version >= LooseVersion('3.6') and driver_version < LooseVersion('3.6'):
-        msg += 'you must use pymongo 3.6+ with MongoDB >= 3.6'
-        module.fail_json(msg=msg)
-
-    elif srv_version >= LooseVersion('3.4') and driver_version < LooseVersion('3.4'):
-        msg += 'you must use pymongo 3.4+ with MongoDB >= 3.4'
-        module.fail_json(msg=msg)
-
-    elif srv_version >= LooseVersion('3.2') and driver_version < LooseVersion('3.2'):
-        msg += 'you must use pymongo 3.2+ with MongoDB >= 3.2'
-        module.fail_json(msg=msg)
-
-    elif srv_version >= LooseVersion('3.0') and driver_version <= LooseVersion('2.8'):
-        msg += 'you must use pymongo 2.8+ with MongoDB 3.0'
-        module.fail_json(msg=msg)
-
-    elif srv_version >= LooseVersion('2.6') and driver_version <= LooseVersion('2.7'):
-        msg += 'you must use pymongo 2.7+ with MongoDB 2.6'
-        module.fail_json(msg=msg)
+from ansible.module_utils.common import check_compatibility
 
 
 def replicaset_status(client, module):
